Conflicts:
	go/src/koding/go-webserver/go-webserver.go
	go/src/koding/go-webserver/templates/loggedin.go

+++ go/src/koding/kites/kloud/provider/amazon/client.go
@@ -114,8 +114,9 @@ func (a *AmazonClient) Build(withPush bool, start, finish int) (artifactResp *pr
 	}
 
 	return &protocol.Artifact{
-		IpAddress:  instance.PublicIpAddress,
-		InstanceId: instance.InstanceId,
+		IpAddress:    instance.PublicIpAddress,
+		InstanceId:   instance.InstanceId,
+		InstanceType: a.Builder.InstanceType,
 	}, nil
 }
 
@@ -190,8 +191,9 @@ func (a *AmazonClient) Start(withPush bool) (*protocol.Artifact, error) {
 	}
 
 	return &protocol.Artifact{
-		InstanceId: instance.InstanceId,
-		IpAddress:  instance.PublicIpAddress,
+		InstanceId:   instance.InstanceId,
+		IpAddress:    instance.PublicIpAddress,
+		InstanceType: a.Builder.InstanceType,
 	}, nil
 }
 
@@ -266,6 +268,10 @@ func (a *AmazonClient) Restart(withPush bool) error {
 }
 
 func (a *AmazonClient) Destroy(start, finish int) error {
+	if a.Id() == "" {
+		return errors.New("instance id is empty")
+	}
+
 	a.Push("Terminating machine", start, machinestate.Terminating)
 	_, err := a.Client.TerminateInstances([]string{a.Id()})
 	if err != nil {
